+++ hardhat.config.ts
@@ -3,10 +3,14 @@
 
 import '@nomicfoundation/hardhat-toolbox';
 import '@openzeppelin/hardhat-upgrades';
+import 'hardhat-dependency-compiler';
 import { HardhatUserConfig } from 'hardhat/config';
 
 const config: HardhatUserConfig = {
     solidity: '0.8.24',
+    dependencyCompiler: {
+        paths: ['@openzeppelin/contracts/proxy/beacon/UpgradeableBeacon.sol'],
+    },
 };
 
 export default config;